+++ test/e2e/user_workload_monitoring_test.go
@@ -88,8 +88,8 @@ func TestUserWorkloadMonitoringMetrics(t *testing.T) {
 			f:    assertTenancyForMetrics,
 		},
 		{
-			name: "assert tenancy model is enforced for labels",
-			f:    assertTenancyForLabels,
+			name: "assert tenancy model is enforced for series metadata",
+			f:    assertTenancyForSeriesMetadata,
 		},
 		{
 			name: "assert prometheus is not deployed in user namespace",
@@ -891,7 +891,7 @@ func assertTenancyForRules(t *testing.T) {
 	}
 }
 
-func assertTenancyForLabels(t *testing.T) {
+func assertTenancyForSeriesMetadata(t *testing.T) {
 	const testAccount = "test-labels"
 
 	err := framework.Poll(2*time.Second, 10*time.Second, func() error {
@@ -974,7 +974,61 @@ func assertTenancyForLabels(t *testing.T) {
 		t.Fatalf("failed to query labels from Thanos querier: %v", err)
 	}
 
-	// Check that /api/v1/label/namespace/values returns a single value.
+	// Check the /api/v1/series endpoint.
+	err = framework.Poll(5*time.Second, time.Minute, func() error {
+		// The tenancy port (9092) is only exposed in-cluster so we need to use
+		// port forwarding to access kube-rbac-proxy.
+		host, cleanUp, err := f.ForwardPort(t, "thanos-querier", 9092)
+		if err != nil {
+			return err
+		}
+		defer cleanUp()
+
+		client := framework.NewPrometheusClient(
+			host,
+			token,
+			&framework.QueryParameterInjector{
+				Name:  "namespace",
+				Value: userWorkloadTestNs,
+			},
+		)
+
+		resp, err := client.Do("GET", "/api/v1/series?match[]=up", nil)
+		if err != nil {
+			return err
+		}
+		defer resp.Body.Close()
+
+		b, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+
+		if resp.StatusCode != http.StatusOK {
+			return fmt.Errorf("unexpected status code response, want %d, got %d (%s)", http.StatusOK, resp.StatusCode, framework.ClampMax(b))
+		}
+
+		res, err := gabs.ParseJSON(b)
+		if err != nil {
+			return err
+		}
+
+		series, err := res.Path("data").Children()
+		if err != nil {
+			return err
+		}
+
+		if len(series) != 1 {
+			return errors.Errorf("expecting a series list with one item, got %d (%s)", len(series), framework.ClampMax(b))
+		}
+
+		return nil
+	})
+	if err != nil {
+		t.Fatalf("failed to query series from Thanos querier: %v", err)
+	}
+
+	// Check that /api/v1/label/{namespace}/values returns a single value.
 	err = framework.Poll(5*time.Second, time.Minute, func() error {
 		// The tenancy port (9092) is only exposed in-cluster so we need to use
 		// port forwarding to access kube-rbac-proxy.